moved model register to admin.py
